+++ idea/testData/refactoring/rename/automaticRenamerOverloadsAmbiguousImport/after/usage/ktUsage.kt
@@ -1,0 +1,9 @@
+package usage
+
+import lib.foo
+import lib.bar
+
+fun test() {
+    bar(1)
+    foo(1, "2")
+}